+++ src/Amazon.S3/Model/S3GetObjectRequest.m
@@ -91,13 +91,13 @@
 -(NSString *)getRange
 {
     if (rangeSet) {
-        return [NSString stringWithFormat:@"bytes=%d-%d", rangeStart, rangeEnd];
+        return [NSString stringWithFormat:@"bytes=%lld-%lld", rangeStart, rangeEnd];
     }
 
     return nil;
 }
 
--(void)setRangeStart:(int)start rangeEnd:(int)end
+-(void)setRangeStart:(int64_t)start rangeEnd:(int64_t)end
 {
     if (end <= start) {
         @throw [AmazonClientException exceptionWithName : @"Invalid range" reason : @"rangeEnd must be larger than rangeStart" userInfo : nil];
