+++ src/Amazon.S3/Model/S3InitiateMultipartUploadRequest.m
@@ -30,13 +30,13 @@
     return self;
 }
 
--(void)setExpires:(int)exp
+-(void)setExpires:(NSInteger)exp
 {
     expires    = exp;
     expiresSet = YES;
 }
 
--(int)expires
+-(NSInteger)expires
 {
     return expires;
 }
