+++ src/include/S3/S3InitiateMultipartUploadRequest.h
@@ -20,11 +20,11 @@
  *
  */
 @interface S3InitiateMultipartUploadRequest:S3AbstractPutRequest {
-    NSString *cacheControl;
-    NSString *contentDisposition;
-    NSString *contentEncoding;
-    int      expires;
-    bool     expiresSet;
+    NSString  *cacheControl;
+    NSString  *contentDisposition;
+    NSString  *contentEncoding;
+    NSInteger expires;
+    bool      expiresSet;
 }
 
 -(id)initWithKey:(NSString *)aKey inBucket:(NSString *)aBucket;
