+++ src/include/S3/S3ListMultipartUploadsRequest.h
@@ -21,13 +21,13 @@
  *
  */
 @interface S3ListMultipartUploadsRequest:S3Request {
-    NSString *delimiter;
-    NSString *keyMarker;
-    NSString *prefix;
-    NSString *uploadIdMarker;
+    NSString  *delimiter;
+    NSString  *keyMarker;
+    NSString  *prefix;
+    NSString  *uploadIdMarker;
 
-    int      maxUploads;
-    bool     maxUploadsIsSet;
+    NSInteger maxUploads;
+    bool      maxUploadsIsSet;
 }
 
 
@@ -71,6 +71,6 @@
 @property (nonatomic, retain) NSString *uploadIdMarker;
 
 /** Gets and sets the maximum number of multipart uploads to return in the response. */
-@property (nonatomic) int maxUploads;
+@property (nonatomic) NSInteger maxUploads;
 
 @end
