+++ src/include/S3/S3ListMultipartUploadsResult.h
@@ -30,7 +30,7 @@
     NSMutableArray *uploads;
 
     bool           isTruncated;
-    int            maxUploads;
+    NSInteger      maxUploads;
 }
 
 
@@ -62,7 +62,7 @@
 @property (nonatomic, readonly) NSMutableArray *uploads;
 
 /** The maxUploads that was specified in the request. */
-@property (nonatomic) int maxUploads;
+@property (nonatomic) NSInteger maxUploads;
 
 /** Indicates whether the returned list of multipart uploads is truncated. */
 @property (nonatomic) bool isTruncated;
