+++ src/include/S3/S3ListPartsRequest.h
@@ -21,9 +21,9 @@
  *
  */
 @interface S3ListPartsRequest:S3Request {
-    NSString *uploadId;
-    int      maxParts;
-    int      partNumberMarker;
+    NSString  *uploadId;
+    NSInteger maxParts;
+    NSInteger partNumberMarker;
 }
 
 /** The uploadId of the multipart upload to list */
