+++ src/include/S3/S3Request.h
@@ -25,14 +25,14 @@
  *
  */
 @interface S3Request:AmazonServiceRequest {
-    NSString *authorization;
-    int      contentLength;
-    NSString *contentType;
-    NSDate   *date;
-    NSString *securityToken;
-    NSString *bucket;
-    NSString *key;
-    NSString *subResource;
+    NSString  *authorization;
+    int64_t   contentLength;
+    NSString  *contentType;
+    NSDate    *date;
+    NSString  *securityToken;
+    NSString  *bucket;
+    NSString  *key;
+    NSString  *subResource;
 }
 
 #pragma mark Properties
@@ -41,7 +41,7 @@
 @property (nonatomic, retain) NSString *authorization;
 
 /** Length of the message (without the headers) according to RFC 2616. */
-@property (nonatomic) int contentLength;
+@property (nonatomic) int64_t contentLength;
 
 /** The content type of the resource.
 * Example: <code>text/plain</code> */
