+++ src/include/S3/S3Response.h
@@ -25,7 +25,7 @@
  *
  */
 @interface S3Response:AmazonServiceResponse {
-    int                 contentLength;
+    int64_t             contentLength;
     NSString            *connectionState;
     NSDate              *date;
     NSString            *etag;
@@ -72,7 +72,7 @@
 @property (nonatomic) bool deleteMarker;
 
 /** The length in bytes of the body in the response. */
-@property (nonatomic) int contentLength;
+@property (nonatomic) int64_t contentLength;
 
 /** Sets the value in the response corresponding to the HTTP header */
 -(void)setValue:(id)value forHTTPHeaderField:(NSString *)header;
